# documentation/templates/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AppHub API",
    "description": "AppHub API",
    "contact": {
      "email": "lsjustin89@gmail.com"
    },
    "license": {
      "name": "GNU General Public License v3.0",
      "url": "https://github.com/myapphub/apphub/blob/main/LICENSE"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "{{ url }}"
    }
  ],
...
  "paths": {
...
    "/user/login": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "user login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserLoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid input, object invalid"
          }
        }
      }
    },
...
    "/user": {
      "get": {
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get my user info",
        "responses": {
          "200": {
            "description": "Get my user info.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update my user info",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "My user info updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid input, object invalid"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "an existing username already exists"
          }
        }
      }
    },
...
    "/user/{provider}/login": {
      "get": {
        "tags": [
          "integration"
        ],
        "summary": "Get auth url",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "enum": [
                "feishu", "slack", "dingtalk", "wecom"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "get auth url",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialAuthUrlResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "integration"
        ],
        "summary": "login",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "enum": [
                "feishu", "slack", "dingtalk"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "user login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialLoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid input, object invalid"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "UserLoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "required": true
          }
        }
      },
      "UserLoginResponse": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "email_verified": {
            "type": "boolean"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "format": "url"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "SocialAuthUrlResponse": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "url"
          }
        }
      },
      "SocialLoginRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "access_token": {
            "type": "string"
          },
          "id_token": {
            "type": "string"
          }
        }
      },
      "SocialLoginResponse": {

      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "email_verified": {
            "type": "boolean"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "format": "url"
          }
        }
      },
      "UserUpdateRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          }
        }
      },
...
    }
  }
}